requirements.txt
Pillow==10.4.0
selenium==4.25.0
webdriver_manager==4.0.2
